pcache/victim: Fix victim cache eviction bug
It is kind of a major rewrite. I learned some lesson from coding pcache.
The refcount is usd carefully.

SMP is really a lot pain in ass.

+++ include/processor/pcache_types.h
@@ -100,7 +100,7 @@ struct pcache_set {
  * new field needs to be initialized in init_pcache_meta_map().
  */
 struct pcache_meta {
-	u8			bits;
+	unsigned long		bits;
 	struct list_head	rmap;
 	atomic_t		mapcount;
 	atomic_t		_refcount;
